reinstate hooks.client.ts
